requirements.txt
bleach==6.1.0
Django==5.1
django-rest-knox==5.0.1
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==28.1.0
tinycss2==1.3.0
